Move modules from HologramImage classs to misc/holography
